submissions/34-plaid-verification/requirements.txt
plaid-python>=14.0.0
python-dotenv>=1.0.0
